Add clang sanitizers
Run the clang address, thread, and undefined behavior sanitizers on the
test executable.

This code uses the CMake sanitizer integration from
https://github.com/arsenm/sanitizers-cmake.

It builds each sanitizer into a different directory, then runs the tests
in that directory.

On Travis, the COMPILER environment variable is used to build with the
proper compiler for each type of build.